+++ CHANGELOG.md
@@ -7,10 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `app-operator.giantswarm.io/paused: "true"` annotation to newly created
+  App CRs installing apps running in Giant Swarm Management Cluster.
+- Add `config-controller.giantswarm.io/version: "0.0.0"` label to newly created
+  App CRs installing apps running in Giant Swarm Management Cluster.
+
 ### Changed
 
-- Add pause annotation to App CRs installing apps running in Giant Swarm
-  Management Cluster.
 - Update cert apiVersion to v1.
 
 ## [0.3.0] - 2021-01-11