+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2021-01-12
+
 ### Added
 
 - Add `app-operator.giantswarm.io/paused: "true"` annotation to newly created
@@ -47,7 +49,8 @@ allow app CR creation.
 - Add validation webhook that is enabled for app CRs with
 `app-operator.giantswarm.io/version` label value >= `3.0.0`.
 
-[Unreleased]: https://github.com/giantswarm/app-admission-controller/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/giantswarm/app-admission-controller/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/giantswarm/app-admission-controller/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/giantswarm/app-admission-controller/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/giantswarm/app-admission-controller/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/giantswarm/app-admission-controller/releases/tag/v0.1.0